% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\title{Calibration model for predicted ITEs}
\usage{
cal(y0hat, y1hat, y, ind.B = NULL)
}
\arguments{
\item{y0hat}{Predicted outcome under control treatment.}

\item{y1hat}{Predicted outcome under treatment.}

\item{y}{Observed outcome of interest.}

\item{ind.B}{Index for individuals in the treated group (should be a vector
of length y with the indices that belong to treated individuals).}
}
\value{
Calibration model for predicted ITEs
}
\description{
Fits a calibration model based on predicted risk under both
treatments and an observed binary outcome. Note that the ITE predictions
that are evaluated should be for those that were treated (ind.B).
}
\examples{
# generate some  data
n <- 250
x <- rnorm(n) # covariate
trt <- rbinom(n, 1, .5) # treatment
beta <- c(0, 1, 1, -.75) # coefficients
lp <- cbind(1, x, trt, x*trt) \%*\% beta # linear predictor
y <- rbinom(n, 1, plogis(lp)) # simulated y

# model
mod <- glm(y ~ x * trt, family = "binomial")

# predict ite's
control.data <- cbind(1, x, 0, 0)
treated.data <- cbind(1, x, 1, x)
y0hat <- plogis(control.data \%*\% coef(mod))
y1hat <- plogis(treated.data \%*\% coef(mod))

ind.B <- which(trt == 1)
cal(y0hat,y1hat,y,ind.B) # apparent
}
