% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbendelta.R
\name{cbendelta}
\alias{cbendelta}
\title{Matching-based (on deltahat) concordance for benefit}
\usage{
cbendelta(deltahat, y, ind.A, ind.B, nresample = 1, get.all = FALSE)
}
\arguments{
\item{deltahat}{Predicted individualized treatment effects}

\item{y}{Observed outcome of interest}

\item{ind.A}{Index for individuals in the control group (should be a vector
of length y with the indices that belong to control individuals).}

\item{ind.B}{Index for individuals in the treated group (should be a vector
of length y with the indices that belong to treated individuals).}

\item{nresample}{In case of unequal  numbers in the treatment groups, the
concordance estimate is based on repeated matching of the smallest group to
equal sized subsamples of the larger group. nresample denotes the number of
replications of this procedure.}

\item{get.all}{get the c-for-ben for all resamplings.}
}
\value{
The concordance estimate(s) for predicted ITE
}
\description{
This function derives the c-for-benefit based on 1:1 control:treated matches,
with matching based on predicted individualized treatment effect. The main
reference is \url{DOI:10.1016/j.jclinepi.2017.10.021}.
}
\examples{
# generate some  data
n <- 250
x <- rnorm(n) # covariate
trt <- rbinom(n, 1, .5) # treatment
beta <- c(0, 1, 1, -.75) # coefficients
lp <- cbind(1, x, trt, x*trt) \%*\% beta # linear predictor
y <- rbinom(n, 1, plogis(lp)) # simulated y

# model
mod <- glm(y ~ x * trt, family = "binomial")

# predict ite's
control.data <- cbind(1, x, 0, 0)
treated.data <- cbind(1, x, 1, x)
deltahat <- plogis(treated.data \%*\% coef(mod)) -
plogis(control.data \%*\% coef(mod))

# cbendelta
table(trt) # check treatment group sizes
ind.A <- which(trt == 0)
ind.B <- which(trt == 1)
res <- cbendelta(deltahat, y, ind.A, ind.B, nresample = 100, get.all=TRUE)
mean(res) # cbendelta
hist(res); abline(v=mean(res), col="red")
}
