% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbeny0.R
\name{cbeny0}
\alias{cbeny0}
\title{Matching-based (on y0hat) concordance for benefit}
\usage{
cbeny0(
  y0hat,
  y1hat,
  y,
  ind.A,
  ind.B,
  match = "y0hat",
  nresample = 1,
  get.all = FALSE
)
}
\arguments{
\item{y0hat}{Predicted outcome under control treatment.}

\item{y1hat}{Predicted outcome under treatment.}

\item{y}{Observed outcome of interest.}

\item{ind.A}{Index for individuals in the control group (should be a vector
of length y with the indices that belong to control individuals).}

\item{ind.B}{Index for individuals in the treated group (should be a vector
of length y with the indices that belong to treated individuals).}

\item{match}{Currently set to "y0hat" (i.e., match on predicted outcome under
control treatment).}

\item{nresample}{In case of unequal  numbers in the treatment groups, the
concordance estimate is based on repeated matching of the smallest group to
equal sized subsamples of the larger group. nresample denotes the number of
replications of this procedure.}

\item{get.all}{get the c-for-ben for all resamplings.}
}
\value{
The concordance estimate(s) for predicted ITE
}
\description{
Matching-based (on y0hat) concordance for benefit
}
\examples{
# generate some  data
n <- 250
x <- rnorm(n) # covariate
trt <- rbinom(n, 1, .5) # treatment
beta <- c(0, 1, 1, -.75) # coefficients
lp <- cbind(1, x, trt, x*trt) \%*\% beta # linear predictor
y <- rbinom(n, 1, plogis(lp)) # simulated y

# model
mod <- glm(y ~ x * trt, family = "binomial")

# predict ite's
control.data <- cbind(1, x, 0, 0)
treated.data <- cbind(1, x, 1, x)
y0hat <- plogis(control.data \%*\% coef(mod))
y1hat <- plogis(treated.data \%*\% coef(mod))

# cbeny0
table(trt) # check treatment group sizes
ind.A <- which(trt == 0)
ind.B <- which(trt == 1)
res <- cbeny0(y0hat, y1hat, y, ind.A, ind.B, "y0hat",
nresample = 250, get.all=TRUE)
mean(res) # cbeny0
hist(res); abline(v=mean(res), col="red")
}
