% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteval-package.R
\docType{package}
\name{iteval-package}
\alias{iteval}
\alias{iteval-package}
\title{iteval}
\description{
The goal of iteval is to ease implementation of the discrimination and
calibration measures for predicted individualized treatment effect as described
by Hoogland et al.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeroenhoogland/iteval}
  \item Report bugs at \url{https://github.com/jeroenhoogland/iteval/issues}
}

}
\author{
\strong{Maintainer}: Jeroen Hoogland \email{j.hoogland-2@umcutrecht.nl} (\href{https://orcid.org/0000-0002-2397-6052}{ORCID})

}
