% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal}
\alias{cal}
\title{Calibration model for predicted ITEs}
\usage{
cal(y0hat, y1hat, y, trt, p0 = NULL, p1 = NULL, type = "both")
}
\arguments{
\item{y0hat}{Predicted outcome under control treatment.}

\item{y1hat}{Predicted outcome under treatment.}

\item{y}{Observed outcome of interest.}

\item{trt}{Vector with treatment allocations (1 for treated, 0 for control).}

\item{p0}{Estimate of the potential outcome under the control condition based on independent
data (or known truth in a simulation setting).}

\item{p1}{Estimate of the potential outcome under the treated condition based on independent
data (or known truth in a simulation setting).}

\item{type}{Default type="both", indicating that the ITE predictions are evaluated in the
control arm based on an offset (the estimated outcome under the treated condition)
estimated in the treated group and given as p1, and in the treated arm based on an offset
(the estimated outcome under the control condition) as estimated in the control group and
given as p0. Types "control" and "treated" only evaluate in this names arm.}
}
\value{
Calibration model for predicted ITEs
}
\description{
Fits a calibration model based on predicted risk under both
treatments and an observed binary outcome. Note that the ITE predictions
that are evaluated should be for those that were treated (ind.B).
}
\examples{
# generate some  data
set.seed(123)
n <- 500
x <- rnorm(n) # covariate
trt <- rbinom(n, 1, .5) # treatment
beta <- c(0, 1, 1, -.75) # coefficients
lp <- cbind(1, x, trt, x*trt) \%*\% beta # linear predictor
y <- rbinom(n, 1, plogis(lp)) # simulated y

# model
mod <- glm(y ~ x * trt, family = "binomial")

# predict ite's
control.data <- cbind(1, x, 0, 0)
treated.data <- cbind(1, x, 1, x)
y0hat <- plogis(control.data \%*\% coef(mod))
y1hat <- plogis(treated.data \%*\% coef(mod))

# truth probabilities for current sim
p0 <- plogis(control.data \%*\% c(0, 1, 1, -.75))
p1 <- plogis(treated.data \%*\% c(0, 1, 1, -.75))

cal(y0hat,y1hat,y,trt,p0,p1,type="both")
}
